13_Kubernetes-I_Intro_EKS/requirements.txt
fastapi==0.115.5
onnxruntime==1.20.1
numpy==2.1.1
pillow==10.4.0
pydantic==2.9.2
python-fasthtml==0.10.0
shad4fast==1.3.1
pytest==8.3.4
sqlite_minutils==4.0.3
gdown